+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 
 ### Added
+  - *Planting view* - Shift selection of planting intervals with mouse or keyboard.
   - *Field map* − Add button show family or crop color.
   - Basic note taking feature for plantings.
   - Harvest page.